refactor(borrow): move utility functions to sagas.utils
